Refactor CreateModulesExecutionProcessMetricUseCase.kt code style, CreateParallelExecutionRateMetricUseCase.kt code style, BuildInfo.kt code base
