+++ src/main/kotlin/io/github/janbarari/gradle/analytics/metric/modulesexecutionprocess/create/CreateModulesExecutionProcessMetricUseCase.kt
@@ -40,10 +40,9 @@ class CreateModulesExecutionProcessMetricUseCase(
 
         modules.whenEach {
             val tasks = input.executedTasks.filter { it.path.startsWith(path) }
-
             val overallDuration = input.getExecutionDuration().toMillis()
             val moduleParallelDuration = tasks.sumOf { it.getDurationInMillis() }
-            val moduleNonParallelDuration = calculateNonParallelExecutionDuration(tasks)
+            val moduleNonParallelDuration = input.calculateNonParallelExecutionDuration(tasks)
             val moduleParallelRate = (moduleParallelDuration - moduleNonParallelDuration)
                 .toPercentageOf(moduleNonParallelDuration)
             val moduleCoverage = moduleNonParallelDuration.toPercentageOf(overallDuration)
@@ -62,93 +61,6 @@ class CreateModulesExecutionProcessMetricUseCase(
         return ModulesExecutionProcessMetric(
             modules = moduleExecutionProcesses
         )
-    }
-
-    @Suppress("NestedBlockDepth", "ComplexMethod")
-    private fun calculateNonParallelExecutionDuration(executedTasks: Collection<TaskInfo>): Long {
-
-        fun checkIfCanMerge(
-            parallelTaskReport: TaskInfo,
-            linearTask: Map.Entry<Int, Pair<Long, Long>>,
-            linearTasks: HashMap<Int, Pair<Long, Long>>
-        ) {
-
-            if (parallelTaskReport.startedAt <= linearTask.value.second &&
-                parallelTaskReport.finishedAt >= linearTask.value.second
-            ) {
-                var start = linearTask.value.first
-                var end = linearTask.value.second
-                if (parallelTaskReport.startedAt < linearTask.value.first) {
-                    start = parallelTaskReport.startedAt
-                }
-                if (parallelTaskReport.finishedAt > linearTask.value.second) {
-                    end = parallelTaskReport.finishedAt
-                }
-                linearTasks[linearTask.key] = Pair(start, end)
-            }
-
-        }
-
-        val nonParallelDurations = hashMapOf<Int, Pair<Long, Long>>()
-
-        val iterator = executedTasks
-            .sortedBy { task -> task.startedAt }
-            .iterator()
-
-        while (iterator.hasNext()) {
-
-            val executedTask = iterator.next()
-            if (nonParallelDurations.isEmpty()) {
-                nonParallelDurations[nonParallelDurations.size] = Pair(executedTask.startedAt, executedTask.finishedAt)
-                continue
-            }
-
-            var shouldBeAddedItem: Pair<Long, Long>? = null
-            val linearTasksIterator = nonParallelDurations.iterator()
-            while (linearTasksIterator.hasNext()) {
-                val linearTask = linearTasksIterator.next()
-
-                checkIfCanMerge(executedTask, linearTask, nonParallelDurations)
-
-                if (executedTask.startedAt > linearTask.value.first &&
-                    executedTask.finishedAt > linearTask.value.second &&
-                    executedTask.finishedAt > executedTask.startedAt
-                ) {
-                    shouldBeAddedItem = Pair(executedTask.startedAt, executedTask.finishedAt)
-                }
-
-            }
-
-            shouldBeAddedItem?.let { new ->
-                val itr = nonParallelDurations.iterator()
-                while (itr.hasNext()) {
-                    val linearTask = itr.next()
-
-                    if (new.first <= linearTask.value.second &&
-                        new.second >= linearTask.value.second
-                    ) {
-                        var start = linearTask.value.first
-                        var end = linearTask.value.second
-                        if (new.first < linearTask.value.first) {
-                            start = new.first
-                        }
-                        if (new.second > linearTask.value.second) {
-                            end = new.second
-                        }
-                        nonParallelDurations[linearTask.key] = Pair(start, end)
-                    }
-                }
-
-                val biggestLTEnd = nonParallelDurations.toList().maxByOrNull { it.second.second }!!.second.second
-                if (new.first > biggestLTEnd) {
-                    nonParallelDurations[nonParallelDurations.size] = new
-                }
-
-            }
-
-        }
-
-        return nonParallelDurations.toList().sumOf { (it.second.second - it.second.first) }
     }
 
 }
